+++ tests/collections_test.py
@@ -41,7 +41,6 @@ from collections import OrderedDict, defaultdict
 import pytest
 
 import jsonpickle
-from jsonpickle.compat import PY2
 
 __status__ = "Stable"
 __version__ = "1.0.0"
@@ -87,11 +86,6 @@ class C(object):
         self.plain_default[key] = (key, value)
 
     def __hash__(self):
-        # return id(self)
-        # the original __hash__() below means the hash can change after being
-        # stored, which JP does not support on Python 2.
-        if PY2:
-            return id(self)
         return hash(self.v) if hasattr(self, 'v') else id(self)
 
     def __repr__(self):
@@ -112,8 +106,6 @@ class D(object):
         self.plain.add(item)
 
     def __hash__(self):
-        if PY2:
-            return id(self)
         return hash(self.v) if hasattr(self, 'v') else id(self)
 
     def __repr__(self):
@@ -218,21 +210,14 @@ def test_dict_self_cycle():
     plain_keys = list(c1u.plain.keys())
     ordered_keys = list(c1u.plain_ordered.keys())
     default_keys = list(c1u.plain_default.keys())
-    if PY2:
-        value = 67
-    else:
-        value = 42
+    value = 42
     assert value == c1u.plain[plain_keys[0]][0].v
     42 == c1u.plain_ordered[ordered_keys[0]][0].v
     42 == c1u.plain_default[default_keys[0]][0].v
 
     # key c2u
     # succeeds because c2u does not have a cycle to itself
-    if PY2:
-        key = c1u
-    else:
-        key = c2u
-    assert key == c1u.plain[plain_keys[1]][0]
+    assert c2u == c1u.plain[plain_keys[1]][0]
     # succeeds because c2u does not have a cycle to itself
     assert c2u == c1u.plain_ordered[ordered_keys[1]][0]
     assert 67 == c1u.plain_default[default_keys[1]][0].v